+++ pkg/OPI/R/OPI-package.r
@@ -18,9 +18,9 @@
 #' }
 #' @seealso
 #' \itemize{
-#'   \item OPI on perimetry.org: \url{http://perimetry.org/OPI}
-#'   \item Andrew's website: \url{http://people.eng.unimelb.edu.au/aturpin/opi/index.html}
+#'   \item Andrew's website: \url{https://people.eng.unimelb.edu.au/aturpin/opi/index.html}
 #'   \item Github OPI repository: \url{https://github.com/turpinandrew/OPI}
+#'   \item OPI Discourse Forum: \url{https://openperimetry.org/}
 #' }
 #' @author Author and mantainer: Andrew Turpin <\email{aturpin@@unimelb.edu.au}>
 #' @references