+++ pkg/OPI/NEWS.md
@@ -123,3 +123,5 @@
  * Added Display as an OPI client.
  * Fixed cdTodb to stop if given <= 0 input
  * Fixed some color="white" examples. (Color is machine specific)
+ * Rewrite of the Display OPI option.(Version 2.9.4)
+ * Updates to DayDream code. (Version 2.9.4)